+++ notebooks/py_scripts/42_Decoding_Clusterless.py
@@ -225,7 +225,7 @@ ContFragClusterlessClassifier(**model_params["decoding_params"])
 # +
 from non_local_detector.environment import Environment
 
-# Environment?
+# ?Environment
 # -
 
 # ## Decoding